Rename wallet_summary_cache to wallet_snapshot_cache
The module caches WalletSnapshot now, not WalletSummary. Nothing else
changes: same coalescing, same seqlock-epoch invalidation, same tests.
