+++ rust/src/wallet/mod.rs
@@ -8,4 +8,4 @@ pub mod sync;
 pub mod sync_engine;
 pub(crate) mod transparent_receive_cache;
 pub mod voting;
-pub(crate) mod wallet_summary_cache;
+pub(crate) mod wallet_snapshot_cache;
