+++ rust/src/wallet/wallet_snapshot_cache.rs
